+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/SessionManagerTest.kt
@@ -4,10 +4,15 @@
 
 package mozilla.components.browser.session
 
+import mozilla.components.browser.state.action.BrowserAction
+import mozilla.components.browser.state.action.TabListAction
+import mozilla.components.browser.state.action.LastAccessAction
 import mozilla.components.browser.state.state.CustomTabConfig
+import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineSessionState
 import mozilla.components.support.test.any
+import mozilla.components.support.test.argumentCaptor
 import mozilla.components.support.test.mock
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
@@ -17,6 +22,8 @@ import org.junit.Test
 import org.mockito.ArgumentMatchers.anyString
 
 import org.mockito.Mockito.`when`
+import org.mockito.Mockito.calls
+import org.mockito.Mockito.inOrder
 import org.mockito.Mockito.mock
 import org.mockito.Mockito.never
 import org.mockito.Mockito.reset
@@ -1102,5 +1109,27 @@ class SessionManagerTest {
         manager.add(sessions)
 
         verify(observer).onSessionsRestored()
+    }
+
+    @Test
+    fun `WHEN restoring a session THEN dispatch updates to store`() {
+        val store = spy(BrowserStore())
+        val inOrder = inOrder(store)
+        val manager = SessionManager(mock(), store)
+        val session = Session("http://www.mozilla.org")
+        val captor = argumentCaptor<BrowserAction>()
+
+        manager.restore(SessionManager.Snapshot(listOf(
+            SessionManager.Snapshot.Item(
+                session,
+                lastAccess = 123
+            )
+        ), 0))
+
+        inOrder.verify(store, calls(2)).dispatch(captor.capture())
+
+        assertTrue(captor.allValues[0] is TabListAction.RestoreAction)
+        assertTrue(captor.allValues[1] is LastAccessAction.UpdateLastAccessAction)
+        assertEquals(123, store.state.tabs[0].lastAccess)
     }
 }